requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
djitellopy>=2.5.0
python-dotenv
#別途、hailo-plarformから.whlファイルをダウンロードしてインストールしてください。
